refactor: move ecstore rebalance migration helpers
